refactor(ts): convert eslint-plugin-development-internal
