+++ eslint/babel-eslint-plugin-development-internal/tsconfig.json
@@ -1,0 +1,14 @@
+/* This file is automatically generated by scripts/generators/tsconfig.ts */
+{
+  "extends": [
+    "../../tsconfig.base.json",
+    "../../tsconfig.paths.json"
+  ],
+  "include": [
+    "./src/**/*.ts",
+    "./src/**/*.cts",
+    "../../lib/globals.d.ts",
+    "../../scripts/repo-utils/*.d.ts"
+  ],
+  "references": []
+}
